Classes have been moved to the new package
